Trenches/Trenches/requirements.txt
# No external dependencies required 
ecdsa==0.18.0
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0 